Add files via upload
The code for the Arduino, Android App, and Windows Form App. All these programs work together to control the ROV. The Arduino sends the PWM waves to the thrusters and controls the servos. The WFA connects the Arduino via USB and tells the Arduino what to do based on the user's actions. The WFA also opens up a TCP socket for the Android app to connect. The Android app sends its accelerometer data to the WFA via the inter/intra net and the WFA in turn forwards that data as commands over USB to the Arduino.